agnes/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
anthropic>=0.39
python-dotenv>=1.0
